Add hiera_include to the default node stanza
This lets us provide always-include classification information to every
node. The primary use case is to bypass the 5-minute waiting period for
ndoes to be added to the Default group in the PE Console, and
immediately include pe_mcollective. This lets us stand up and snapshot
new stacks faster and more reliably.