Reduce dynacism in configuration
Previously, we used the $environment variable in the hiera_config
directive on the Puppet master. This no longer works in the latest
version(s) of PE, so the path to hiera.yaml is now being hard-coded.
Note that this doesn't actually change much since the hiera_config
variable was not checked at the beginning of every run (only when the
master was loaded) so it didn't actually work to use it in dynamic
environments anyway.

Also switching from setting manifestdir for dynamic environments to
setting manifest, since the latter still works but the same changes that
broke $environment in hiera_config seem to have broken it in
manifestdir.