Segregate tools for building or bootstrapping
This commit creates a new directory "tools" and moves content not
directly related to a Puppet environment into it. This includes one-time
manifests, runtime scripts, and packaging scripts.

The desired outcome of this commit is to clean up the root directory so
that it is clearer what the important aspects of a Puppet control
repository are. Because of the way environment workflows are used it
doesn't make sense to replicate all the build tooling any time the
Puppet environment is forked.

In the future, it may make sense to house the build tools in a separate
repository altogether, or create a Puppet module to install them to a
classified system. The main reason they are bundled here is for
convenience of artifact building, not due to any particular design
benefit.