Add rakefile to define build actions
Previously a simple shell script was used to build the project.

The build process has become complex enough that it makes sense to break
it up into component steps which can be optionally parameterized. The
Rakefile format is a purpose-specific technology that closely aligns
with other tooling in use at Puppet Labs. It provides us the ability to
better define individual build actions, put the logic in a well-known
location, and break apart the individual actions in a way that can
provide an interface to a CI tool like Jenkins in the event that it
makes sense for the CI tool to perform intermediate logic not strictly
part of a simple build.