Refactor deploy.sh script to use new path (/tools)
Previously a directory /scripts was used to house executables being
shipped as part of a Puppet environment. When all collateral, including
non-executable content was merged into a single directory that directory
was renamed to the more generic "tools". This commit updates the
deploy.sh script to reference the new path.