Fix initial one-time Puppet applications
When the *.pp files were moved from the top of the environment into the
tools directory the tools/deploy.sh script was not completely updated to
reflect that move. As such, two one-time puppet apply runs were not made
on the master. This commit fixes that problem so that the one-time
puppet apply runs will install the correct configuration.