Ensure puppet-control uses "production" branch
Previously the branch the generated puppet-control git repository used
would have the same name as the current branch of the working repo. This
is undesirable as the generated control repo needs to represent the
default puppet environment which is always "production".

This commit ensures that the branch name of the generated puppet-control
repo will always be "production".