Update r10k remote parameter
This change is being made in prep for a new build process wherein the
Puppet Enterprise master will have a local copy of all repositories in
/opt/puppet/repos.

This is the next step up. Previously the r10k configuration was laid
down but not used except in unusual circumstances, now the r10k config
will be self-contained and local but fully functional on a new master.

Also remove the empty parameter list from the role to better indicate
that roles should not take parameters.

This seems like it should probably be a profile...