Switch stderr silencer to `git branch -m`
There are two commands in the shell statement, one to try first and one
to try afterwards if the first one fails. This commit moves the
/dev/null silencer from the fallback command to the try-first command.
The end result is that no errors will be shown by the first command,
which has a fallback, but if the fallback errors out those errors will
be shown.