Increase readability of Rakefile
This commit alters the Rakefile by removing many ENV['var'] file
references, replacing them with direct string references to the relative
file paths.

The extra flexibility of using variables to reference the various build
paths is not worth the cognitive cost of reading and understanding what
the Rakefile does when everything is a variable. Using the actual file
paths makes for a much improved experience understanding what is going
on in the file.