Cherry-pick production changes fd9624d..0daf217
This commit cherry-picks the following changes from the production
branch into the 3.8.x branch.

(SET-85) Strip git history when building artifact

This commit causes the built artifact to automatically strip any
existing project history. Previously, if a module had been cloned from
an existing git project, all the history in it would have been retained.
For our offline use case it is preferable to have a smaller artifact
than to retain that much history.

Change default task to only build repos tarball

When running `rake` with no task specified, only build the artifact we
actually use. The code to create a basic tarball is not being removed
yet, though the longer we go without using it the more confidence we
have that we can remove the logic from the Rakefile.

Migrate from opentable/ modules to puppet/ modules

The opentable modules are no longer maintained and have been migrated to
puppet-community. This commit updates our Puppetfile to reference the
updated modules. They should be compatible with all of our code as
really all that's happend is the projects have moved.

Update dism, firewall references

Firewall needs to have a specific version pinned for reproducibility.
Dism has updates in Git that have yet to be released to the Forge that
we need in order for it to work with 2015.2.0, so switching temporarily
to git sourcing. We will move back to the Forge when a fixed release is
published.

Update all Forge module references

This commit updates all referenced Forge modules to the latest available
semver-compatible versions. Note that ntp is updated to a new major as
none of the breaking changes listed should affect our usage of it.