(SET-85) Strip git history when building artifact
This commit causes the built artifact to automatically strip any
existing project history. Previously, if a module had been cloned from
an existing git project, all the history in it would have been retained.
For our offline use case it is preferable to have a smaller artifact
than to retain that much history.