Update dism, firewall references
Firewall needs to have a specific version pinned for reproducibility.
Dism has updates in Git that have yet to be released to the Forge that
we need in order for it to work with 2015.2.0, so switching temporarily
to git sourcing. We will move back to the Forge when a fixed release is
published.