(SET-92) Move hiera.yaml to new default location
Previously we configured hiera.yaml in the control repo of production:

    /etc/puppetlabs/code/environments/production/hiera.yaml

This can mislead people into thinking they can have different hiera.yaml
files per environment, which is not the case.

This commit updates the control repo to remove the hiera.yaml, and
ensure that the global hiera.yaml in $codedir is created correctly.