(SET-106) Ensure default color scheme for vim in demo env is elflord
Changed profile::vim to give it a colorscheme parameter and set it's
default to 'elflord' after consensus was reach in Demo Dojo that it
was most legible for screen/projector/colorblind staff.

Changed classifier.sh script to add profile::vim (with explicit
setting of colorscheme param to elflord) to 'PE Master' and 'Linux Servers'
node groups.