(maint) Update to a newer fork of puppet/dotnet
The stable puppet/dotnet release does not account for ensuring some
versions of .NET present on OSes that include them; e.g. .NET 4.0 on
Server 2012 R2. This commit switches to a newer fork of .NET that works
more correctly on the OSes we support. The fork may eventually be merged
into puppet/dotnet, at which point we can switch back to the official
release.