Update firewall module to patched version for EL 7
On CentOS 7 with firewall rule purging, when firewalld is shut down it
removes rules that Puppet has decided to purge mid-run. Without the
patch, these show up as resource errors. (It's apparently not good
enough that the rule is already absent, Puppet gets upset that it didn't
get to remove the rule itself).

This commit updates the Puppetfile to point to a patched version with a
fix. This will allow us to migrate to CentOS 7 without hitting those
errors.