Convert staging.pp: staging::file => remote_file
staging is still a good choice when needing to extract archives that it
supports, but remote_file is a cleaner resource for simple file
fetching. This commit updates the staging.pp file both to use
remote_file, and also with some Puppet 4 language iteration to make
Splunk *slightly* less ugly. (it's still bad).