(maint) Remove unnecessary resource defaults
Leading up to the switch-over from Puppet 3 to Puppet 4 there were
several warnings about resource defaults that would be thrown which we
silenced by setting explicit resource defaults in site.pp to be the
to-be-defaults in the upcoming Puppet 4. Now that Puppet 4 is live, we
don't need to specify these defaults as they are the default defaults.