Convert classifier.sh to classifier.pp
This commit changes the shell script used to fire node group
configuration into the NC to a puppet manifest for the same. It also
adds the prosvcs/node_manager module to the Puppetfile to enable this to
work.

Because of a critical bootstrapping requirement, we're sourcing the
module from git and a commit at this early stage. We should be able to
transition to a Forge release shortly.