Revert to puppet-apply then puppet-agent
We keep too much classification for the Master in the PE Console.
Because of this, it's not elegant to try to fit everything into one
single puppet-apply run. Therefore, minimize the puppet-apply to the
smallest possible bootstrap configuration (set up the node manager),
then run puppet-agent to do the rest.