Import node groups for app orchestration demo
This commit updates the master's node manager configuration to be ready
to run an App Orchestrator demo with the rgbank content.

This is being done to enable demonstration of app orchestrator with the
TSE demo environment when 2015.3.0 launches tomorrow. In the future,
this kind of demo-specific node manager configuration should be
enablable / disablable, rather than baked in to the default config.
There is a SET ticket for that.