Make improvements to Rakefile verbosity
Right now the tar command outputs more than 10,000 lines of log to the
screen whenever a build is performed. That's over the top. This commit
removes the verbose flag from tar and adds --quiet to a few git commands
to get things back under control so it's easier to tell what's going on
(right and wrong) in a build.