Adding RGBank wordpress example.
In this version of RGBank I've updated it slightly to 1) not say it
was deployed via AO, and 2) Combined all the steps for deployment into
one manifest.  The general idea with this is you can show a customer
deploying a custom LAMP stack application end-to-end instead of just
showing a finished wordpress installation.