Fix file permissions on oradb source download
Previously there was a missing path component on the file resource(s)
that ensured the permissions were set correctly on the files downloaded
to the master to support the oradb demo. This commit squashes the extra
file resource into the remote_file resource directly to just set the
mode there, solving the problem and simplifying the code.