Move demo repos tarball to its own class
It turns out that the exec which creates the pe demo repos tarball is
dependent on profile::master::fileserver. Rather than make any part of
profile::master::puppetserver dependent on that, it is cleaner to move
the pe demo configuration to be under p::m::fileserver, and modify the
role to list it just like all of the other fileserver content the puppet
master includes.