Fix persistence bug loading git repos
See comments in source. Basically, we had an exec resource creating a
file, using a creates=> stanza, but the location the file was being
created in was shared persistent storage. This commit updates the logic
to make sure the shared, persistent storage has a symlink installed
pointing back to a local, non-shared build of the data that the master
creates.