Fix the rgbank demo
This patch fixes two things that are broken about the rgbank demo.
First, it locks the vcsrepo code checkout to a specific commit. Without
this, we would not have a valid Wordpress site because the upstream repo
has moved files around a bunch.

Second, it modifies the way the appserver role is deployed to not cause
Puppet errors prior to application deployment. This means the nodes
won't be red when we run `puppet job run`.