Provide `puppet code-commit` command
This is a work-around for being able to use the code-staging directory
directly, until PE-15438 and CODEMGMT-697 are solved.

A new command is placed in the /opt/puppetlabs/bin directory, which is
in the root user's PATH. The command can be invoked as follows:

    puppet code-commit

There is a single variation:

    puppet code-commit -v

The simple invocation commits all code in the code-staging directory and
deploys it. The -v makes the command print some messaging about what
it's doing as it does it.