Add a deploy complete signal file to the master
This file will be placed in the webroot of the apache instance that the
master runs, and can be used as a signal to external, unauthenticated
watchers that the puppet master is completely deployed.

We need this kind of signal for deploying the stack in Slice, or AWS.
Right now the Slice openstack deployment doesn't let us use native
watchers, so we have to implement script hacks ourselves.