Switch away from Forge module for selinux
Because of cached catalog use. The Forge module makes use of facts to
dynamically build resources into the catalog based on whether or not it
is necessary to disable SELinux. That doesn't work so well if the first
catalog built is cached, because then it will forever more be trying to
run `setenforce permissive`.