(maint) Fix permissions on deployed user accounts
At some point the names of the RBAC permissions we try to give users
changed. That shoulda been fine except that even though there's no API
version number bump if we try to assign the old names for the
permissions the server explodes and we get no perms.

This commit fixes it by removing the no-longer-valid change lifetime
perm, and changing the name of the orchestrator perm to the modern
equivalent.