Update master fileserver sources to use fact, port
Port 81 is now used for the master's fileserver, so we need to update
sources that reference it. While doing this, might as well stop
hard-coding the master's name as a string, and update it to use the
$::puppet_server fact which resolves to whatever name/address the agent
is configured to use to reach the master.

This requires an updated version of stdlib.